+++ R/utils_expand_data_to_grid.R
@@ -5,17 +5,23 @@
 #' @importFrom purrr map map_lgl map_df modify_if
 #' @importFrom sjlabelled as_numeric
 #' @importFrom dplyr n_distinct
+#' @importFrom tidyselect ends_with
 # fac.typical indicates if factors should be held constant or not
 # need to be false for computing std.error for merMod objects
-get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, type = "fe", prettify = TRUE, prettify.at = 25, pretty.message = TRUE) {
+get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, type = "fe", prettify = TRUE, prettify.at = 25, pretty.message = TRUE, condition = NULL) {
   # special handling for coxph
   if (inherits(model, "coxph")) mf <- dplyr::select(mf, -1)
 
   # make sure we don't have arrays as variables
-  mf <- purrr::modify_if(mf, is.array, as.vector)
+  mf <- suppressWarnings(purrr::modify_if(mf, is.array, as.vector))
 
   # use tibble, no drop = FALSE
   mf <- tibble::as_tibble(mf)
+
+  # check for logical variables, might not work
+  if (any(purrr::map_lgl(mf, is.logical))) {
+    stop("Variables of type 'logical' do not work, please coerce to factor and fit the model again.", call. = FALSE)
+  }
 
   # # make sure we don't have arrays as variables
   # mf[, 2:ncol(mf)] <- purrr::modify_if(mf[, 2:ncol(mf)], is.array, as.vector)
@@ -25,57 +31,6 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
   w <- get_model_weights(model)
   if (all(w == 1)) w <- NULL
 
-  ## TODO handle random effects in predict correctly
-
-  # The problem is that I’m not sure how to define the random effects for
-  # the "newdata"-argument, see this example:
-  #
-  # library(lme4)
-  # library(ggplot2)
-  #
-  # data("sleepstudy")
-  #
-  # sleepstudy$age <- round(rnorm(nrow(sleepstudy), mean = 50, sd = 15))
-  # sleepstudy$education <- factor(sample(c("low", "medium", "high"), size = nrow(sleepstudy), replace = T))
-  #
-  # m <- lmer(Reaction ~ Days + age + education + (1 + Days | Subject), data = sleepstudy)
-  #
-  # dat <- data.frame(expand.grid(
-  #   Days = unique(sleepstudy$Days),
-  #   age = mean(sleepstudy$age),
-  #   education = levels(sleepstudy$education)[1],
-  #   Subject = unique(sleepstudy$Subject)
-  # ))
-  #
-  # dat$predicted1 <- predict(m, newdata = dat, re.form = NULL, type = "response")
-  # dat$predicted2 <- predict(m, newdata = dat, re.form = NA, type = "response")
-  #
-  # ggplot(dat, aes(x = Days, y = predicted1)) + geom_line()
-  # ggplot(dat, aes(x = Days, y = predicted2)) + geom_line()
-  #
-  # In the above example, ignoring the random effects in "predict()" gives a
-  # straight line, while accounting for them results in a "zigzag"-line -
-  # but this only happens when providing a "newdata"-argument.
-  #
-  # So, what 'ggpredict()' currently does, is using the first method (the zig-zag-line),
-  # but the random effects in "newdata" are set to just one level (the first
-  # level in the random effect variable).
-  #
-  # If anyone has a solution to this, how to incorporate all levels of random
-  # effects, let me know.
-  # get random effects, if any
-
-  rand.eff <- NULL
-  # tryCatch(
-  #   {
-  #     rand.eff <- names(lme4::ranef(model))
-  #   },
-  #   error = function(x) { NULL },
-  #   warning = function(x) { NULL },
-  #   finally = function(x) { NULL }
-  # )
-
-
   # clean variable names
   colnames(mf) <- sjstats::var_names(colnames(mf))
 
@@ -84,6 +39,29 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
   # and all specified variables
   rest <- get_clear_vars(terms)
 
+
+  # check if user has any predictors with log-transformatio inside
+  # model formula, but *not* used back-transformation "exp". Tell user
+  # so she's aware of the problem
+
+  tryCatch(
+    {
+      if (!inherits(model, "brmsfit") && pretty.message) {
+        log.terms <- grepl("^log\\(([^,)]*).*", x = attr(stats::terms(model), "term.labels", exact = TRUE))
+        if (any(log.terms)) {
+          clean.term <- sjstats::pred_vars(model)[which(log.terms)]
+          exp.term <- tidyselect::ends_with("[exp]", vars = terms)
+
+          if (sjmisc::is_empty(exp.term) || get_clear_vars(terms)[exp.term] != clean.term) {
+            message(sprintf("Model has log-transformed predictors. Consider using `terms = \"%s [exp]\"` to back-transform scale.", clean.term))
+          }
+        }
+      }
+    },
+    error = function(x) { NULL },
+    warning = function(x) { NULL },
+    finally = function(x) { NULL }
+  )
 
   # create unique combinations
   rest <- rest[!(rest %in% names(first))]
@@ -126,11 +104,6 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
   # keep those, which we did not process yet
   alle <- alle[!(alle %in% names(first))]
 
-  # exclude random effects
-  if (!is.null(rand.eff) && type == "re")
-    alle <- alle[!(alle %in% rand.eff)]
-
-
   # if we have weights, and typical value is mean, use weighted means
   # as function for the typical values
 
@@ -139,6 +112,15 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
 
   if (typ.fun == "weighted.mean" && sjmisc::is_empty(w))
     typ.fun <- "mean"
+
+
+  # do we have variables that should be held constant at a
+  # specific value?
+
+  if (!is.null(condition) && !is.null(names(condition))) {
+    first <- c(first, as.list(condition))
+    alle <- alle[!(alle %in% names(condition))]
+  }
 
 
   # add all to list. For those predictors that have to be held constant,
@@ -163,16 +145,11 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
   # add constant values.
   first <- c(first, const.values)
 
-  # add back random effects
-  if (!is.null(rand.eff) && type == "re")
-    first <- c(first, lapply(mf[, rand.eff], unique))
-
-
   # reduce and prettify elements with too many values
   if (prettify) {
     .pred <- function(p) is.numeric(p) && dplyr::n_distinct(p) > prettify.at
     too.many <- purrr::map_lgl(first, .pred)
-    first <- purrr::map_if(first, .p = .pred, ~ pretty(.x, n = round(sqrt(diff(range(.x))))))
+    first <- purrr::map_if(first, .p = .pred, pretty_range)
 
     if (any(too.many) && pretty.message) {
       message(sprintf(