+++ NEWS.md
@@ -1,3 +1,33 @@
+# ggeffects 0.4.0
+
+## General
+
+* Removed alias names `mem()`, `eff()` and `ame()`.
+* For mixed models (packages **lme4**, **nlme**, **glmmTMB**), the uncertainty of the random effect variances is now taken into account when `type = "re"`.
+* Computing confidence intervals for mixed models should be much more memory efficient now, resulting less often in warnings about memory allocation problems.
+* Updated reference in `CITATION` to the publication in the Journal of Open Source Software.
+* A test-suite was added to the package.
+
+## New functions
+
+* `pretty_range()`, to create a pretty sequence of integers of a vector.
+
+## Changes to functions
+
+* `ggpredict()` gets a `condition`-argument to specify values at which covariates should be held constant, instead of their `typical` value.
+* The `pretty`-option for `ggpredict()` now calculates more values, leading to smoother plots.
+* The `terms`-argument in `ggpredict()` can now also select a range of feasible values for numeric values, e.g. `terms = "age [pretty]"`. In contrast to the `pretty`-argument, which prettyfies all terms, you can selectively prettify specific terms with this option.
+* The `terms`-argument in `ggpredict()` now also supports all shortcuts that are possible for the `mdrt.values`-argument in `gginteraction()`, so for instance `term = "age [meansd]"` would return three values: mean(age) - sd(age), mean(age) and mean(age) + sd(age).
+* `plot()` gets some new arguments to control which plot-title to show or hide: `show.title`, `show.x.title` and `show.y.title`.
+* `plot()` gets a `log.y` argument to transform the y-axis to logarithmic scale, which might be useful for binomial models with predicted probabilities, or other models with log-alike link-functions.
+* The `plot()`-method for plotting all effects with `ggpredict()` (when `term = NULL`) now allows to arrange the plot in facets (using `facets = TRUE`).
+* Values in dot-argument for `plot()` are now passed down to `ggplot::scale_y*()`, to control the appearance of the y-axis (like `breaks`).
+
+## Bug fixes
+
+* Fixed issue with binomial models that used `cbind(...)` as response variable.
+* Fixed issue with suboptimal precision of confidence resp. prediction intervals for mixed models (packages **lme4**, **nlme**), which are now more accurate.
+
 # ggeffects 0.3.4
 
 ## General